Add an _XML method implementation to the XML instance for Foundation.String
The XML instance for Foundation.String has until this patch been missing
an implementation of the `_XML` method. Aside from the shortcomings in
functionality, this causes a build error using the default configuration
with more recent (8.2.1+) versions of GHC:

  src/Text/XML/Hexml/Lens.hs:102:10: error: [-Wmissing-methods, -Werror=missing-methods]
      • No explicit implementation for
          ‘_XML’
      • In the instance declaration for ‘XML F.String’
      |
  102 | instance XML F.String where
      |          ^^^^^^^^^^^^

In prior versions of GHC, -Wmissing-methods would not warn on missing type
class methods having names beginning with an underscore. This behavior was
changed in ghc/ghc@503219e

The Iso between Foundation.String and Strict.ByteString introduced by this
patch performs a copy when converting to a ByteString, as I did not find a
straightforward implementation without one.